+++ src/Raven.CodeAnalysis/TypeInfo.cs
@@ -3,12 +3,22 @@ namespace Raven.CodeAnalysis;
 public class TypeInfo
 {
     internal TypeInfo(ITypeSymbol? type, ITypeSymbol? convertedType, Conversion conversion = default)
+        : this(type, convertedType, conversion, type, convertedType)
+    {
+    }
+
+    internal TypeInfo(
+        ITypeSymbol? type,
+        ITypeSymbol? convertedType,
+        Conversion conversion,
+        ITypeSymbol? flowType,
+        ITypeSymbol? convertedFlowType)
     {
         Type = type;
         ConvertedType = convertedType;
         Conversion = conversion;
-        Nullability = CreateNullabilityInfo(type);
-        ConvertedNullability = CreateNullabilityInfo(convertedType);
+        Nullability = CreateNullabilityInfo(type, flowType);
+        ConvertedNullability = CreateNullabilityInfo(convertedType, convertedFlowType);
     }
 
     public Conversion Conversion { get; }
@@ -21,14 +31,18 @@ public class TypeInfo
 
     public ITypeSymbol? Type { get; }
 
-    private static NullabilityInfo CreateNullabilityInfo(ITypeSymbol? typeSymbol)
+    private static NullabilityInfo CreateNullabilityInfo(ITypeSymbol? typeSymbol, ITypeSymbol? flowType)
     {
         if (typeSymbol is null)
             return new NullabilityInfo(NullableAnnotation.None, NullableFlowState.None);
 
-        if (typeSymbol.IsNullable)
-            return new NullabilityInfo(NullableAnnotation.Annotated, NullableFlowState.MaybeNull);
+        var annotation = typeSymbol.IsNullable
+            ? NullableAnnotation.Annotated
+            : NullableAnnotation.NotAnnotated;
+        var flowState = flowType is null
+            ? typeSymbol.IsNullable ? NullableFlowState.MaybeNull : NullableFlowState.NotNull
+            : flowType.IsNullable ? NullableFlowState.MaybeNull : NullableFlowState.NotNull;
 
-        return new NullabilityInfo(NullableAnnotation.NotAnnotated, NullableFlowState.NotNull);
+        return new NullabilityInfo(annotation, flowState);
     }
 }